joblistings-backend/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
SQLAlchemy
mysql-connector-python
python-dotenv
selenium